+++ src/Symfony/Component/Cache/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 5.4
 ---
 
+ * Make `LockRegistry` use semaphores when possible
  * Deprecate `DoctrineProvider` because this class has been added to the `doctrine/cache` package
 
 5.3